+++ NEWS.md
@@ -1,18 +1,49 @@
 > Copyright © 2016 RTE Réseau de transport d’électricité
 
-
-# antaresRead 2.5.3
+# antaresRead 2.7.0
 
 ### Breaking changes (Antares v8.7) : 
 
 * Existing function `readBindingConstraints()` (cf. Antares v8.7 changelog)
 
-# antaresRead 2.5.2
 
-BREAKING CHANGES:
+# antaresRead 2.6.1 (devlopment)
+
+BUGFIXES :  
+
+* `setSimulationPathAPI()` :  
+  - returns an API exception if the requested study ID is incorrect 
+  - `simulation` the simulation parameter works with negative values within the limit of the number of simulations
+* correction in `readClusterDesc()` calls to add "opts"
+
+
+# antaresRead 2.6.0
+
+NEW FEATURES (Antares v8.6, cf. Antares v8.6 changelog) :
+
+* `readClusterSTDesc()` read "short-term storage" clusters parameters (input files of an antares study)
+
+
+BREAKING CHANGES (Antares v8.6) :
+
+* `readInputTS()` is now compatible to read time series with :  
+  - "short-term storage"  
+  - "mingen" (pmin hydro value)
+* `setSimulationPath()` has new parameter `areasWithSTClusters` (name of area with "st-storage" cluster)
+
+
+BUGFIXES : 
 
 * `setSimulationPathAPI` generate new global parameter `sleep` to add timer to API request
+* Correction of `.importOutput()` to use `readAntares()` with `parallel == TRUE` in shiny application
+* `setSimulationPathAPI()` delete a redundant API request
+* `readClusterDesc()` minor fix in API mode + fix if no cluster exists => return specific error message
+* `readIniAPI()` read well file `generaldata` for sections "playlist" and "variables selection"
 
+
+DATA : 
+
+* A test study in tar.gz format is available in version `v8.6.0`
 
 
 # antaresRead 2.5.1